Squash a large number of warnings in the tests (#3447)
* Add explicit observed=False to groupby calls per upcoming pandas change

* Remove deprecated approach to setting contour label, which never actually worked

* Remove test case that raises seaborn deprecation warning

* Silence pandas deprecation warning

* Avoid pandas warning for setting nan on int serires

* Addressed contour API change missed in #3446

* Only cast integer series to floats before adding nulls

* Avoid pandas SettingWithCopyWarning

* Address another case of setting nan on int series

* Handle upcoming deprecation of pd.unique(list)

* Avoid casting nonfinite values to intp while property mapping

* Ignore numpy warning when all violins are singular

* Catch another invalid cast

* Ignore persistent xdist warning from test machinery

* Don't convert warnigns to errors in test config

+++ tests/_stats/test_density.py
@@ -195,7 +195,7 @@ class TestKDE:
 
         val, ori = "xy"
         df["weight"] = 1
-        df = df[[ori, "weight"]]
+        df = df[[ori, "weight"]].astype(float)
         df.loc[:4, col] = np.nan
         gb = self.get_groupby(df, ori)
         res = KDE()(df, gb, ori, {})
